requirements.txt
flask
flask-cors
flask-jwt-extended
flask-limiter
flask-talisman
boto3
pinecone-client
langchain
jinja2
python-dotenv
marshmallow
openai
redis
cryptography
bcrypt
pycryptodome
pyjwt
python-multipart
werkzeug
validators
bleach
